축제 스켈레톤 파일을 스켈레톤 폴더로 이동,스타일 수정
