+++ src/Components/Product/ProductStyle.js
@@ -58,7 +58,6 @@ export const ProductBox = styled.li`
     font-weight: 400;
   }
 `;
-
 export const Nav = styled.nav`
   display: flex;
   margin: 10px auto;
@@ -74,38 +73,5 @@ export const Nav = styled.nav`
   @media (min-width: 768px) {
     max-width: 480px;
     margin-top: 0px;
-  }
-`;
-export const SkeletonContainer = styled(ProductContainer)`
-  gap: 20px;
-  margin: 10px auto;
-  margin-left: 10px;
-  .itemImage {
-    border-radius: 10px;
-    width: 370px;
-    height: 180px;
-    margin-bottom: 5px;
-    @media (min-width: 768px) {
-      width: 280px;
-    }
-  }
-
-  .itemName {
-    margin: 0 auto;
-    margin: 13px 0 6px 4px;
-    width: 150px;
-    height: 15px;
-    @media (min-width: 768px) {
-    }
-  }
-  .itemDate {
-    margin-left: 4px;
-    width: 140px;
-  }
-  @media (min-width: 768px) {
-    grid-column-gap: 43px;
-    grid-row-gap: 20px;
-    max-width: 480px;
-    margin: 0px auto;
   }
 `;
